String hash in a separated file
