+++ SobrassadaEngine/ResourceManagement/Resources/ResourceStateMachine.h
@@ -1,28 +1,13 @@
 #pragma once
 
 #include "Resource.h"
+#include "HashString.h"
 
 #include "Math/float2.h"
 #include "imgui.h"
 #include <string>
 #include <vector>
 
-struct HashString
-{
-    size_t hash = 0;
-    std::string original;
-
-    HashString() = default;
-    HashString(const std::string& str)
-    {
-        original = str;
-        hash     = std::hash<std::string> {}(str);
-    }
-
-    bool operator==(const HashString& other) const { return hash == other.hash; }
-    bool operator!=(const HashString& other) const { return !(*this == other); }
-    const std::string& GetString() const { return original; }
-};
 
 struct Clip
 {
