+++ SobrassadaEngine/SobrassadaEngine.vcxproj.filters
@@ -1040,6 +1040,9 @@
     <ClInclude Include="Utils\NavMeshConfig.h">
       <Filter>Utils</Filter>
     </ClInclude>
+    <ClInclude Include="Utils\HashString.h">
+      <Filter>Utils</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <None Include="Libs\MathGeoLib\include\Geometry\TriangleMesh_IntersectRay_CPP.inl">
